Refactor: 500억 이상 거래대금 별찍기

+++ project/stock_helper/src/main/java/com/example/stock_helper/signalEvening/ItemForm.java
@@ -4,7 +4,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 @Data
 public class ItemForm {
-    private int stockRise;//상승률
+    private float stockRise;//상승률
     private long stockTransactionAmount;//거래대금
     private MultipartFile attachFile;//이브닝리포트
     private boolean writeDate;//날짜표기할껀지
